requirements.txt
img2pdf
Pillow
tkinterdnd2
